\name{gxgRC-package}
\alias{gxgRC-package}
\alias{gxgRC-pkg}
\docType{package}
\title{
\packageTitle{gxgRC}
}
\description{
\packageDescription{gxgRC}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{gxgRC}
\packageIndices{gxgRC}
%~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{gxgRC}

Maintainer: \packageMaintainer{gxgRC}
}
\references{
%~~ Literature or other references for background information ~~
}
%~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
%~~ Optional links to other man pages, e.g. ~~
%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%~~ simple examples of the most important functions ~~
}
